requirements.txt
qrcode==6.1
XlsxWriter==1.1.8
Django==3.1.7
pandas
fpdf2==2.3.5
mysql_connector_repackaged==0.3.1
secrets
